Spring Security / CH Connect configuration
Add Spring Security and configure it with the MITRE OpenID Connect Client library pointing to CH Connect. This setup is suited mainly for development; for production use, we want to use a more secure configuration with signed or encrypted requests, static client configuration and a proper authorities mapper (depends on WISVCH/connect#4).